+++ GRDB/Core/Database.swift
@@ -841,11 +841,13 @@ extension Database {
             self.rawValue = rawValue
         }
         
+        #if !swift(>=4.1)
         /// The hash value
         /// :nodoc:
         public var hashValue: Int {
             return rawValue.hashValue
         }
+        #endif
         
         /// The `BINARY` built-in SQL collation
         public static let binary = CollationName("BINARY")
@@ -878,11 +880,13 @@ extension Database {
             self.rawValue = rawValue
         }
         
+        #if !swift(>=4.1)
         /// The hash value
         /// :nodoc:
         public var hashValue: Int {
             return rawValue.hashValue
         }
+        #endif
         
         /// The `TEXT` SQL column type
         public static let text = ColumnType("TEXT")