+++ GRDB/Core/DatabaseFunction.swift
@@ -5,12 +5,27 @@
 #endif
 
 /// An SQL function or aggregate.
-public final class DatabaseFunction {
-    public let name: String
-    let argumentCount: Int32?
+public final class DatabaseFunction: Hashable {
+    // SQLite identifies functions by (name + argument count)
+    private struct Identity: Hashable {
+        let name: String
+        let nArg: Int32 // -1 for variadic functions
+        
+        #if !swift(>=4.1)
+        var hashValue: Int {
+            return name.hashValue ^ nArg.hashValue
+        }
+        
+        static func == (lhs: Identity, rhs: Identity) -> Bool {
+            return lhs.name == rhs.name && lhs.nArg == rhs.nArg
+        }
+        #endif
+    }
+    
+    public var name: String { return identity.name }
+    private let identity: Identity
     let pure: Bool
     private let kind: Kind
-    private var nArg: Int32 { return argumentCount ?? -1 }
     private var eTextRep: Int32 { return (SQLITE_UTF8 | (pure ? SQLITE_DETERMINISTIC : 0)) }
     
     /// Returns an SQL function.
@@ -38,8 +53,7 @@ public final class DatabaseFunction {
     ///       exactly *argumentCount* elements, provided *argumentCount* is
     ///       not nil.
     public init(_ name: String, argumentCount: Int32? = nil, pure: Bool = false, function: @escaping ([DatabaseValue]) throws -> DatabaseValueConvertible?) {
-        self.name = name
-        self.argumentCount = argumentCount
+        self.identity = Identity(name: name, nArg: argumentCount ?? -1)
         self.pure = pure
         self.kind = .function{ (argc, argv) in
             let arguments = (0..<Int(argc)).map { index in
@@ -89,8 +103,7 @@ public final class DatabaseFunction {
     ///       have exactly *argumentCount* elements, provided *argumentCount* is
     ///       not nil.
     public init<Aggregate: DatabaseAggregate>(_ name: String, argumentCount: Int32? = nil, pure: Bool = false, aggregate: Aggregate.Type) {
-        self.name = name
-        self.argumentCount = argumentCount
+        self.identity = Identity(name: name, nArg: argumentCount ?? -1)
         self.pure = pure
         self.kind = .aggregate { return Aggregate() }
     }
@@ -104,8 +117,8 @@ public final class DatabaseFunction {
         
         let code = sqlite3_create_function_v2(
             db.sqliteConnection,
-            name,
-            nArg,
+            identity.name,
+            identity.nArg,
             eTextRep,
             definitionP,
             kind.xFunc,
@@ -127,8 +140,8 @@ public final class DatabaseFunction {
     func uninstall(in db: Database) {
         let code = sqlite3_create_function_v2(
             db.sqliteConnection,
-            name,
-            nArg,
+            identity.name,
+            identity.nArg,
             eTextRep,
             nil, nil, nil, nil, nil)
         
@@ -274,7 +287,7 @@ public final class DatabaseFunction {
             let aggregateContextU = Unmanaged.passRetained(aggregateContext)
             var aggregateContextP = aggregateContextU.toOpaque()
             withUnsafeBytes(of: &aggregateContextP) {
-                aggregateContextBufferP.copyBytes(from: $0)
+                aggregateContextBufferP.copyMemory(from: $0)
             }
             return aggregateContextU
         }
@@ -309,17 +322,17 @@ public final class DatabaseFunction {
     }
 }
 
-extension DatabaseFunction : Hashable {
+extension DatabaseFunction {
     /// The hash value
     /// :nodoc:
     public var hashValue: Int {
-        return name.hashValue ^ nArg.hashValue
+        return identity.hashValue
     }
     
     /// Two functions are equal if they share the same name and arity.
     /// :nodoc:
     public static func == (lhs: DatabaseFunction, rhs: DatabaseFunction) -> Bool {
-        return lhs.name == rhs.name && lhs.nArg == rhs.nArg
+        return lhs.identity == rhs.identity
     }
 }